+++ .eslintrc.js
@@ -16,5 +16,6 @@ module.exports = {
   plugins: ['@typescript-eslint'],
   rules: {
     '@typescript-eslint/ban-ts-comment': 1,
+    'no-unreachable': 'error'
   },
 }
