+++ packages/client/src/pages/error/error.tsx
@@ -1,23 +1,29 @@
 import React from 'react'
-import { useRouteError } from 'react-router-dom'
+import { isRouteErrorResponse, useRouteError } from 'react-router-dom'
 
 const ErrorPage = () => {
   const error = useRouteError()
-  console.error(error)
-
-  return (
-    <div id="error-page">
-      <h1>Oops!</h1>
-      <p>Sorry, an unexpected error has occurred.</p>
-      <p>
-        <b>{error.status}</b>
-      </p>
-      <p>
-        <i>{error.statusText || error.message}</i>
-      </p>
-    </div>
-  )
-  return <div>ErrorPage</div>
+  if (isRouteErrorResponse(error)) {
+    return (
+      <div id="error-page">
+        <h1>Oops!</h1>
+        <p>Sorry, an unexpected error has occurred.</p>
+        <p>
+          <b>{error.status}</b>
+        </p>
+        <p>
+          <i>{error.statusText}</i>
+        </p>
+      </div>
+    )
+  } else {
+    return (
+      <div>
+        <h1>Oops!</h1>
+        <p>Sorry, an unexpected error has occurred.</p>
+      </div>
+    )
+  }
 }
 
 export default ErrorPage
